Remove stance leg service/client and get stance leg from updated state_estimator/state

+++ ros2/src/march_gait_planning/include/march_gait_planning/gait_planning.hpp
@@ -13,7 +13,7 @@ public:
     explicit GaitPlanning();
 
     // Setters
-    void setStanceFoot(const int &new_stance_foot);
+    void setStanceFoot(const uint8_t &new_stance_foot);
     void setFootPositions(const std::array<double, 3> &new_left_foot_position, const std::array<double, 3> &new_right_foot_position);
     void setGaitType(const exoState &new_gait_type);  
     void setBezierGait();
